🧹 Deduplicate zod/devtools typing surfaces
Share record guards through @umpire/core/guards and reuse devtools extension types from @umpire/devtools so zod avoids shadow type copies while keeping runtime dependencies unchanged.
