+++ packages/zod/package.json
@@ -42,6 +42,7 @@
     "access": "public"
   },
   "devDependencies": {
+    "@umpire/devtools": "workspace:^",
     "zod": "^3.23.8"
   }
 }
